Add tests for configurable TIME-WAIT reuse delay
This set of tests accompanies the kernel patch set which adds a knob for
setting the TIME-WAIT reuse delay [1].

Tests cover connection reincarnation through TIME-WAIT reuse with the
default reuse delay (1 second) as well as with a reuse delay radically
shortened down to 1 millisecond.

We exercise both the happy and failure scenarios, where TW reuse is not
possible because the required delay period has not elapsed yet.

[1] <link TBD>